Part 2/src/Django/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2
filelock==3.11.0
platformdirs==3.2.0
sqlparse==0.4.3
virtualenv==20.21.0
boto3==1.26.142
PyJWT==2.7.0
